Refactor out MerchantPage component
